token

+++ routes/api.php
@@ -11,14 +11,14 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 Route::post('/register', [AuthController::class, 'registrarse']);
 Route::post('/login', [AuthController::class, 'login']);
-Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
+Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
 Route::get('/mostrar', [AuthController::class, 'mostrar']);
-Route::get('/isAdmin/{id}', [AuthController::class, 'isAdmin']);
+Route::get('/isAdmin/{id}', [AuthController::class, 'isAdmin'])->middleware('auth:sanctum');
 
 Route::prefix('/personas')->group(function() {
-    Route::post('/', [PersonasController::class, 'agregar']);
-    Route::put('/{id}', [PersonasController::class, 'modificar']);
-    Route::delete('/{id}', [PersonasController::class, 'eliminar']);
+    Route::post('/', [PersonasController::class, 'agregar'])->middleware('auth:sanctum');
+    Route::put('/{id}', [PersonasController::class, 'modificar'])->middleware('auth:sanctum');
+    Route::delete('/{id}', [PersonasController::class, 'eliminar'])->middleware('auth:sanctum');
     Route::get('/', [PersonasController::class, 'mostrar']);
     Route::get('mostrar/{id}', [PersonasController::class, 'mostrarUnico']);
 });
